requirements.txt
flask
flask-rq # called as flask_rq
flask-caching
asyncio
flask[Async]
requests
python-dotenv # called as dotenv
redis
rq